Updated prod workflow versions and template for tn and umccrise
Upgraded versions to allow for bam to not be generated in germline directory.    

This requires a umccrise upgrade to copy over the normal bam to the umccrise directory first.

We also update the tn template to not generate the germline bam AND to also not run the sv caller (which requires the map align output to be present).